internal: add CString and ErrorWithLog
